+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - [](https://github.com/vegaprotocol/protos/pull/) -
 
 ### 🛠 Improvements
+- [418](https://github.com/vegaprotocol/protos/issues/418) - Add support for cursor based pagination in `Data Node API V2`
 - [](https://github.com/vegaprotocol/protos/pull/) -
 
 ### 🐛 Fixes